Fix for merging of vars to successive for loop
When the variable defined in a declarator is referenced only within the
for loop that proceeds it then it can be safely put into for loop's
init. If it's referenced elsewhere, then only vars can be merged, and
this merging will change the scope for let and const.

This fixes the issue and verifies that there are no other references
apart from the ones within the loop for let/const and doesn't check
this for vars.

+ (close #198)