Add option: topLevel to Mangler
Adds an option topLevel to mangler which defaults to `false` and when
specified to be true, mangles topLevel/global variables too.

Fixes an bug, or a maybe bug, with scope.hasOwnBinding check. Since we
traverse the program once more within the mangler to make it a separate
pass, the outer most scope is not traversed by this traversal, and
variables declared in the global/toplevel belong to this scope - like
Functions, Vars, etc... and will never pass scope.hasOwnBinding test.
So, in the traversals outermost scope we check the binding in its own
scope (for classes) and its parent scope (or mangler.program.scope)
(for functions and variables).

+ (close #204)