Fix bug: local variables don't see toplevel renames
Toplevel scope rename is special and requires renaming in the
mangler.program.scope too instead of just scopable - This is becuase,
as mentioned in the previous commit, the mangler.program.traverse's
first Scopable visitor's scope is not the same as the
mangler.program.scope itself.

We do this rename not based on tracking scope, but based on the binding
itself - we check if a binding is a global/toplevel and rename in
mangler.program.scope if it is so.

Also, we check the program scope for the binding when it is the top
level scope of the traverse - i.e. scope.parent ===
mangler.program.scope.