Fix bug: support for for..of/for..await in DCE
for..of and for..await loops are not considered in DCE and unused var
declarations in these loops are removed and this results in an empty
LVal. The same exception which is considered for a for..in loop is
also applied for for..of or for..await loops, and we use an Alias
that Babel provides - ForXStatement which is a for..in, for..of, or
a for..await loop.

+ (Close #207)