Require blocks around single block-scope declaration (#220)
* Require blocks around single block-scope declaration

A single block scoped declaration under an if/else requires a block
statement. The block statement here if removed is a syntax error and
this removal of block statement is valid only for vars and not for let
or const.

We could essentially go ahead and remove the left hand side of the
variable declaration, but that is the job of the DeadCode elimination
plugin. Here, we just correct the mistake and let simplify generate
correct output when used in isolation. Let dead code take care of
removal of block and the left hand side of the declaration

* Add function declaration to needs block