Transform property literal reserved keywords as described in the documentation.
The documentation describes that when using babel-plugin-transform-property-literals, reserved keywords used as identifiers should have quotes around them, however this does not happen. There are FIXME mentions in the test files, but this is not mentioned in the documentation. Hence I've fixed the code and tests to align with the documentation