Add missing babel-helper-evaluate-path dependency
At the moment, the `babel-plugin-minify-guarded-expressions` package is calling `require('babel-helper-evaluate-path')` without actually depending on this package. It causes an exception to be thrown when the npm does not flatten the dependency tree. You can trigger it by forcing npm to install packages without flattening (like it did in npm@2) by running: npm install --legacy-bundling